requirements.txt
Flask_HTTPAuth==4.1.0
Flask_Negotiate==0.1.0
Flask==0.12.2
waitress==1.4.4
Flask_Cors==3.0.3
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
python_bcrypt==0.3.2
werkzeug==0.16.1
uuid
datetime
json